// src/app/models/Enums/TicketStatus.enum.ts
export enum TicketStatus {
  TO_DO = 'TO_DO',
  PROGRESS = 'PROGRESS',
  OPEN = 'OPEN',
  CLOSED = 'CLOSED',
  REOPENED = 'REOPENED',
  RESOLVED = 'RESOLVED'
}

export enum TicketStatusLabel {
  TO_DO = 'To do',
  PROGRESS = 'Progress',
  OPEN = 'Open',
  CLOSED = 'Closed',
  REOPENED = 'Reopened',
  RESOLVED = 'Resolved'
}